Unwrap most `unittest` test cases to bare functions
