+++ tests/messages/consts.py
@@ -10,3 +10,7 @@ data_dir = os.path.join(this_dir, 'data')
 project_dir = os.path.join(data_dir, 'project')
 i18n_dir = os.path.join(project_dir, 'i18n')
 pot_file = os.path.join(i18n_dir, 'temp.pot')
+
+
+def get_po_file_path(locale):
+    return os.path.join(i18n_dir, locale, 'LC_MESSAGES', 'messages.po')
