+++ tests/test_plural.py
@@ -10,7 +10,6 @@
 # individuals. For the exact contribution history, see the revision
 # history and logs, available at https://github.com/python-babel/babel/commits/master/.
 import decimal
-import unittest
 
 import pytest
 
@@ -198,76 +197,24 @@ def test_tokenize_malformed(rule_text):
         plural.tokenize_rule(rule_text)
 
 
-class TestNextTokenTestCase(unittest.TestCase):
-
-    def test_empty(self):
-        assert not plural.test_next_token([], '')
-
-    def test_type_ok_and_no_value(self):
-        assert plural.test_next_token([('word', 'and')], 'word')
-
-    def test_type_ok_and_not_value(self):
-        assert not plural.test_next_token([('word', 'and')], 'word', 'or')
-
-    def test_type_ok_and_value_ok(self):
-        assert plural.test_next_token([('word', 'and')], 'word', 'and')
-
-    def test_type_not_ok_and_value_ok(self):
-        assert not plural.test_next_token([('abc', 'and')], 'word', 'and')
+def test_next_token_empty():
+    assert not plural.test_next_token([], '')
 
 
-def make_range_list(*values):
-    ranges = []
-    for v in values:
-        if isinstance(v, int):
-            val_node = plural.value_node(v)
-            ranges.append((val_node, val_node))
-        else:
-            assert isinstance(v, tuple)
-            ranges.append((plural.value_node(v[0]),
-                           plural.value_node(v[1])))
-    return plural.range_list_node(ranges)
+def test_next_token_type_ok_and_no_value():
+    assert plural.test_next_token([('word', 'and')], 'word')
 
 
-class PluralRuleParserTestCase(unittest.TestCase):
+def test_next_token_type_ok_and_not_value():
+    assert not plural.test_next_token([('word', 'and')], 'word', 'or')
 
-    def setUp(self):
-        self.n = plural.ident_node('n')
 
-    def n_eq(self, v):
-        return 'relation', ('in', self.n, make_range_list(v))
+def test_next_token_type_ok_and_value_ok():
+    assert plural.test_next_token([('word', 'and')], 'word', 'and')
 
-    def test_error_when_unexpected_end(self):
-        with pytest.raises(plural.RuleError):
-            plural._Parser('n =')
 
-    def test_eq_relation(self):
-        assert plural._Parser('n = 1').ast == self.n_eq(1)
-
-    def test_in_range_relation(self):
-        assert plural._Parser('n = 2..4').ast == \
-            ('relation', ('in', self.n, make_range_list((2, 4))))
-
-    def test_negate(self):
-        assert plural._Parser('n != 1').ast == plural.negate(self.n_eq(1))
-
-    def test_or(self):
-        assert plural._Parser('n = 1 or n = 2').ast ==\
-            ('or', (self.n_eq(1), self.n_eq(2)))
-
-    def test_and(self):
-        assert plural._Parser('n = 1 and n = 2').ast ==\
-            ('and', (self.n_eq(1), self.n_eq(2)))
-
-    def test_or_and(self):
-        assert plural._Parser('n = 0 or n != 1 and n % 100 = 1..19').ast == \
-            ('or', (self.n_eq(0),
-                    ('and', (plural.negate(self.n_eq(1)),
-                             ('relation', ('in',
-                                           ('mod', (self.n,
-                                                    plural.value_node(100))),
-                                           (make_range_list((1, 19))))))),
-                    ))
+def test_next_token_type_not_ok_and_value_ok():
+    assert not plural.test_next_token([('abc', 'and')], 'word', 'and')
 
 
 EXTRACT_OPERANDS_TESTS = (
